move dashboard n add dynamic row creation
