+++ lib/pages/dashboard/dashboard_con.dart
@@ -1,0 +1,7 @@
+import 'package:a_check_web/pages/dashboard/dashboard.dart';
+import 'package:flutter/material.dart';
+
+class DashboardState extends State<Dashboard> {
+  @override
+  Widget build(BuildContext context) => DashboardView(this);
+}
